Add a test suite for sidecar cmd and signals
Add test cover for:

* long-running and short-lived cmd and cmd_args invocations
* signalling via pid_file_name and renew_signal
* windows cmd and cmd_args execution

If rebased onto the 'fix-windows' branch for windows test fixes, these
pass for windows test runs too.

To make these tests possible the following changes are made to the
sidecar code:

* Sidecar.certReadyChan yields *workloadapi.X509Context not struct{}
  so the context can be observed. This could also be useful for
  code extending spiffe-helper.
* A new Sidecar.cmdExitChan channel reports whenever a process
  managed by spiffe-helper exits, and the exit state. This allows
  tests to wait for process exits and see the exit code, etc.
* A new Sidecar.pidFileSignalledChan channel reports whenever
  signalling the file in pid_file_name is attempted, the pid
  signalled, and the outcome. This allows signal based tests
  to see what spiffe-helper tried to signal.

Channels with values are used to ensure that there's no race between
what might be present in the Sidecar struct state when the test observes
it and what was there when the event of interest occurred.

It might be handy to have a channel for observing when the managed
process is started too, but the tests just busy-wait polling
Sidecar.processRunning to detect this for now.

I considered combining these channels into some kind of unified observer
channel that emits a union struct depending on the event type, but it
seemed pointless to do this when we can just select{} on multiple
channels of interest and keep each channel simple.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>